refactor(dapp-request): consolidate approval flow into new feature dir
