+++ src/components/Transactions/constants.ts
@@ -2,9 +2,9 @@ import { type TransitionConfig } from 'moti';
 import { Easing } from 'react-native-reanimated';
 
 import { type EventInfo } from '@/components/Transactions/types';
+import { RequestSource } from '@/features/dapp-request/types';
 import * as i18n from '@/languages';
 import { Screens } from '@/state/performance/operations';
-import { RequestSource } from '@/utils/requestNavigationHandlers';
 import safeAreaInsetValues from '@/utils/safeAreaInsetValues';
 
 export const SCREEN_BOTTOM_INSET = safeAreaInsetValues.bottom + 20;
