+++ src/features/dapp-request/hooks/useTransactionSetup.ts
@@ -6,7 +6,8 @@ import type useGas from '@/features/gas/hooks/useGas';
 import * as i18n from '@/languages';
 import { type ChainId } from '@/state/backendNetworks/types';
 import { methodRegistryLookupAndParse } from '@/utils/methodRegistry';
-import { type RequestSource } from '@/utils/requestNavigationHandlers';
+
+import { type RequestSource } from '../types';
 
 type TransactionSetupParams = {
   chainId: ChainId;
