--- src/parsers/__mocks__/requests.js
@@ -1,3 +1,0 @@
-import { jest } from '@jest/globals';
-
-export const getRequestDisplayDetails = jest.fn();
